requirements.txt
streamlit==1.32.0
Pillow==10.2.0
requests==2.31.0
python-dotenv==1.0.1
tensorflow==2.16.1
opencv-python==4.9.0.80
plotly==5.18.0 